Refactor fcConnection into fcConnectionService
