+++ src/services/index.ts
@@ -6,9 +6,10 @@ import configuration from './configuration';
 export {ForceService, operatingSystem, configuration};
 import DXService from './dxService';
 export const dxService = DXService.getInstance();
-import { FCConnectionService, FCOauth } from './fcConnection';
+import { FCOauth, FCConnection } from './fcConnection';
+import { FCConnectionService } from './fcConnectionService';
 export const fcConnection = FCConnectionService.getInstance();
-export { FCOauth };
+export { FCOauth, FCConnectionService, FCConnection };
 import { CommandViewService } from './commandView';
 export const commandViewService = CommandViewService.getInstance();
 export { Task } from './commandView';
